Add files via upload
This task is a case of Multiclass image classification where the model predicts one of the digits from 0 to 9 to which the input image belongs.

In the MNIST digit recognition task, we use a CNN network to develop a model to recognize the handwritten digit. We would download the MNIST dataset which consists of a training set of 60000 images and 10000 images for testing. Each image is cropped into 28x28 pixels and handwritten digits are from 0 to 9.